web/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
gunicorn==20.1.0
pandas==1.3.0
plotly==5.3.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tenacity==8.2.2
Werkzeug==2.2.3
